Feature/change derived field editor

See merge request BI/magreport2.1!182

+++ magreport-backend/src/main/java/ru/magnit/magreportbackend/service/ReportService.java
@@ -10,6 +10,8 @@ import ru.magnit.magreportbackend.domain.user.SystemRoles;
 import ru.magnit.magreportbackend.dto.inner.RoleView;
 import ru.magnit.magreportbackend.dto.inner.filter.FilterRequestData;
 import ru.magnit.magreportbackend.dto.request.ChangeParentFolderRequest;
+import ru.magnit.magreportbackend.dto.request.filterreport.FilterAddRequest;
+import ru.magnit.magreportbackend.dto.request.filterreport.FilterGroupAddRequest;
 import ru.magnit.magreportbackend.dto.request.folder.CopyFolderRequest;
 import ru.magnit.magreportbackend.dto.request.folder.FolderAddRequest;
 import ru.magnit.magreportbackend.dto.request.folder.FolderChangeParentRequest;
@@ -51,9 +53,11 @@ import ru.magnit.magreportbackend.service.domain.SecurityFilterDomainService;
 import ru.magnit.magreportbackend.service.domain.UserDomainService;
 import ru.magnit.magreportbackend.service.jobengine.filter.FilterQueryExecutor;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -223,15 +227,33 @@ public class ReportService {
 
         var uniqueNames = new HashSet<String>();
         var duplicateNames = new HashSet<String>();
+        var invalidFieldNames = new HashSet<String>();
+        var invalidFilterNames = new HashSet<String>();
 
         request.getFields().forEach(field -> {
             var name = field.getName().toUpperCase();
             if (!uniqueNames.add(name)) duplicateNames.add(field.getName());
+
+            if (Boolean.FALSE.equals(dataSetDomainService.checkSyncDatasetField(field.getDataSetFieldId()))) {
+                invalidFieldNames.add(name);
+            }
         });
 
-        if (!duplicateNames.isEmpty()) {
+        if (!duplicateNames.isEmpty())
             throw new InvalidParametersException("Отчет содержит повторяющиеся имена полей: " + duplicateNames);
-        }
+        if (!invalidFieldNames.isEmpty())
+            throw new InvalidParametersException("Отчет содержит невалидные поля: " + invalidFieldNames);
+
+        unWind(request.getFilterGroup(), new ArrayList<>()).stream()
+                .flatMap(f -> f.getFields().stream())
+                .filter(f -> Objects.nonNull(f.getReportFieldId()))
+                .forEach(f -> {
+                    if (Boolean.FALSE.equals(reportDomainService.checkValidField(f.getReportFieldId())))
+                        invalidFilterNames.add(f.getName());
+                });
+
+        if (!invalidFilterNames.isEmpty())
+            throw new InvalidParametersException("Отчет содержит невалидные поля фильтров: " + invalidFilterNames);
 
         filterReportDomainService.removeFilters(request.getId());
         reportDomainService.deleteFields(reportDomainService.getDeletedFields(request));
@@ -277,7 +299,7 @@ public class ReportService {
         return reportDomainService.changeParentFolder(request);
     }
 
-    public void setReportEncrypt(ReportEncryptRequest request){
+    public void setReportEncrypt(ReportEncryptRequest request) {
         reportDomainService.setReportEncrypt(request);
     }
 
@@ -349,5 +371,13 @@ public class ReportService {
 
         var newFolders = reportDomainService.copyReportFolder(request, userDomainService.getCurrentUser());
         return newFolders.stream().map(f -> reportDomainService.getFolder(userDomainService.getCurrentUser().getId(), f)).toList();
+    }
+
+    private List<FilterAddRequest> unWind(FilterGroupAddRequest reportFilterGroupData, List<FilterAddRequest> reportFilterData) {
+
+        if (reportFilterGroupData == null) return reportFilterData;
+        reportFilterData.addAll(reportFilterGroupData.getFilters());
+        reportFilterData.addAll(reportFilterGroupData.getChildGroups().stream().flatMap(group -> unWind(group, reportFilterData).stream()).toList());
+        return reportFilterData;
     }
 }
